Change[Org hook]: From olivetti to writeroom
Olivetti is a good mode for writing. Writeroom is also a good mode for
writing. Both have things I enjoy, but also have small things I would
prefer to change. Olivetti tries to format your text in a certain way
that looks acceptable for written prose. The problem here is that a
lot of times the writing I do is mixed with code blocks which olivetti
formats badly. Writeroom on the other hand does not attempt at forcing
any sort of formatting, which is not necessarily what I want, but at
least the code blocks are left alone.