hack[org-pomodoro]: nonexistent notifications without icons...
I always use emacs on full screen, *always*! So it was quite annoying
to find that my pomodoro notifications stopped appearing over emacs on
full screen when I updated ubuntu to a more recent version.
The hack I found around this was to change the notification function
used on the pomodoro package into one that has the urgency cranked up
to maximum.
The other problem was the icon and name of the app... it was showing
"unknown application" and once again I just specified the name trough
the :app-name parameter and the correct path to the emacs icon trough
the :app-icon parameter.